+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Change `compare-ddbj` to always count elements as multiset
 * Change to accept case variants in feature/qualifier names
 * Add TOPOLOGY feature if Is_circular=true exists in "source"
+* Change exons to be merged with their parent of type *_segment and *_regionas well as RNA.
 
 
 ## v0.3.0 (2021-10-12)